pyproject.toml
[project]
name = "geo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
